Move api mocks to correct path
