+++ Makefile
@@ -41,7 +41,7 @@ help: ## Display this help.
 .PHONY: gen-ias-client
 gen-ias-client: ## Generate IAS client and client mocks from OpenAPI spec
 	oapi-codegen --package=api -generate=client,types -o ./internal/ias/internal/api/ias.gen.go ./internal/ias/internal/api/SCI_Application_Directory.yaml
-	mockery --name=ClientWithResponsesInterface --dir=./internal/ias/internal/api --output=./internal/ias/internal/mocks --outpkg=mocks --case=underscore
+	mockery --name=ClientWithResponsesInterface --dir=./internal/ias/internal/api --output=./internal/ias/internal/api/mocks --outpkg=mocks --case=underscore
 
 
 .PHONY: manifests
